Move bucket creation from authentication/apps.py to consultations/app.py
